Test_offsets moved to test_date_offsets
