+++ pandas/tests/tseries/offsets/test_yqm_offsets.py
@@ -24,7 +24,7 @@ from pandas.tseries.offsets import (
 )
 
 from .common import assert_offset_equal, assert_onOffset
-from .test_offsets import Base
+from .test_date_offsets import Base
 
 # --------------------------------------------------------------------
 # Misc
